fix(usage): fix CI errors and clamp floating window position
- enable macos-private-api for transparent window on macOS

- remove unnecessary i32 casts in settings command for Clippy

- clamp restored position against available monitors to prevent off-screen window

- add unit tests for clampRestoredPosition

+++ src-tauri/src/commands/settings.rs
@@ -219,8 +219,8 @@ pub fn get_floating_work_area(window: tauri::Window) -> Result<Option<FloatingWo
             let p = monitor.position();
             let s = monitor.size();
             Ok(Some(FloatingWorkArea {
-                x: p.x as i32,
-                y: p.y as i32,
+                x: p.x,
+                y: p.y,
                 width: s.width as i32,
                 height: s.height as i32,
             }))
